+++ hyoga/open/__init__.py
@@ -5,7 +5,9 @@
 Hyoga input tools to open glacier modelling datasets.
 """
 
+from .example import example
 from .local import dataset, mfdataset, subdataset
 
 __all__ = [
+    'example',
     'dataset', 'mfdataset', 'subdataset']